restructure tests using jest describe
