+++ src/parser.ts
@@ -3,7 +3,7 @@
  * to update HTML for next-book specific use cases.
  * @module
  */
-import { ParsedObj, Ideas } from './structures';
+import { ParsedObj, Ideas, IdeaPiece } from './structures';
 import { Delimiter } from './config';
 
 export class Separator {}
@@ -20,7 +20,7 @@ export class Separator {}
 export default function parse(node: Node, delimiter: Delimiter): ParsedObj {
   if (!node) throw new Error('Expected node undefined.');
   if (typeof delimiter == 'undefined') throw new Error('Expected delimiter undefined.');
-  const pieces: (Node | string | Separator | ParsedObj)[] = [];
+  const pieces: IdeaPiece[] = [];
   // first create a flat list of strings, HTML Elements, ParsedObjs, and Separators
   node.childNodes.forEach(child => {
     if (child.nodeType === child.TEXT_NODE) {
@@ -45,6 +45,7 @@ export default function parse(node: Node, delimiter: Delimiter): ParsedObj {
 
   pieces.forEach(piece => {
     if (piece instanceof Separator) ideas.addIdea();
+    // is it anchor or not
     else if (piece instanceof ParsedObj) ideas.addObj(piece);
     else ideas.appendToIdea(piece);
   });
